+++ framework/src/Volo.Abp.BlazoriseUI/AbpBlazoriseModule.cs
@@ -1,4 +1,5 @@
 ﻿using Blazorise;
+using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp.AspNetCore.Components.WebAssembly;
 using Volo.Abp.Modularity;
 
@@ -18,6 +19,8 @@ namespace Volo.Abp.BlazoriseUI
         {
             context.Services
                 .AddBlazorise();
+
+            context.Services.AddSingleton(typeof(AbpBlazorMessageLocalizerHelper<>));
         }
     }
 }
