+++ modules/identity/src/Volo.Abp.Identity.Blazor/AbpIdentityBlazorModule.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Volo.Abp.AspNetCore.Components.WebAssembly.Theming.Routing;
 using Volo.Abp.AutoMapper;
+using Volo.Abp.BlazoriseUI;
 using Volo.Abp.Modularity;
 using Volo.Abp.PermissionManagement.Blazor;
 using Volo.Abp.UI.Navigation;
@@ -10,7 +11,8 @@ namespace Volo.Abp.Identity.Blazor
     [DependsOn(
         typeof(AbpIdentityHttpApiClientModule),
         typeof(AbpAutoMapperModule),
-        typeof(AbpPermissionManagementBlazorModule)
+        typeof(AbpPermissionManagementBlazorModule),
+        typeof(AbpBlazoriseUIModule)
         )]
     public class AbpIdentityBlazorModule : AbpModule
     {
@@ -32,8 +34,6 @@ namespace Volo.Abp.Identity.Blazor
             {
                 options.AdditionalAssemblies.Add(typeof(AbpIdentityBlazorModule).Assembly);
             });
-
-            context.Services.AddSingleton(typeof(AbpIdentityBlazorMessageLocalizerHelper<>));
         }
     }
 }
